fixed bug in extract_SWR
When trying to remove events that occur within a given window (say argument InterRippleInterval =1 which removes swrs that co-occur within 1 sec.), the code was getting stuck. This was fixed by adding a line that only runs the while loop if there were more than 1 swr event in a trial (while loop was getting stuck on these exceptions).